+++ core/src/com/biglybt/core/download/impl/DownloadManagerAvailabilityImpl.java
@@ -192,7 +192,7 @@ DownloadManagerAvailabilityImpl
 			    			}
 
 			    			@Override
-						    public int
+						    public long
 			    			getUploadSpeedKBSec(
 			    				boolean	estimate )
 			    			{
