+++ core/src/com/biglybt/core/lws/LightWeightSeed.java
@@ -667,7 +667,7 @@ LightWeightSeed
 					}
 
 					@Override
-					public int
+					public long
 					getUploadSpeedKBSec(
 						boolean estimate)
 					{
