+++ core/src/com/biglybt/core/speedmanager/impl/v2/PingSpaceMon.java
@@ -206,7 +206,7 @@ public class PingSpaceMon
     {
 
         @Override
-        public int getBytesPerSec() {
+        public long getBytesPerSec() {
             return 1;
         }
